requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.6.0
decorator==5.0.9
gym==0.18.3
idna==2.10
iniconfig==1.1.1
msgpack==1.0.2
numpy==1.21.0
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyglet==1.5.15
git+git://github.com/DanMcInerney/pymetasploit3.git@498d65f1cd132b62071b78df9dc95fa987abf39e#egg=pymetasploit3
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.18.0
requests==2.25.1
retry==0.9.2
scipy==1.7.0
toml==0.10.2
urllib3==1.26.6
